+++ apps/native-component-list/src/screens/MailComposerScreen.tsx
@@ -1,43 +1,11 @@
 import * as MailComposer from 'expo-mail-composer';
 import React from 'react';
-import { Alert, StyleSheet, Text, View } from 'react-native';
+import { Alert, StyleSheet, View } from 'react-native';
 
 import Button from '../components/Button';
 import MonoText from '../components/MonoText';
-import { useResolvedValue } from '../utilities/useResolvedValue';
 
 export default function MailComposerScreen() {
-  const [isAvailable, error] = useResolvedValue(MailComposer.isAvailableAsync);
-
-  const warning = React.useMemo(() => {
-    if (error) {
-      return `An unknown error occurred while checking the API availability: ${error.message}`;
-    } else if (isAvailable === null) {
-      return 'Checking availability...';
-    } else if (isAvailable === false) {
-      // On iOS device without Mail app installed it is possible to show mail composer,
-      // but it isn't possible to send that email either way.
-      return `It's not possible to send an email on this device. Make sure you have mail account configured and Mail app installed (iOS).`;
-    }
-    return null;
-  }, [error, isAvailable]);
-
-  if (warning) {
-    return (
-      <View style={{ justifyContent: 'center', alignItems: 'center', flex: 1 }}>
-        <Text>{warning}</Text>
-      </View>
-    );
-  }
-
-  return <MailComposerView />;
-}
-
-MailComposerScreen.navigationOptions = {
-  title: 'MailComposer',
-};
-
-function MailComposerView() {
   const [status, setStatus] = React.useState<MailComposer.MailComposerStatus | null>(null);
 
   const sendMailAsync = async () => {
@@ -57,11 +25,24 @@ function MailComposerView() {
 
   return (
     <View style={styles.container}>
-      <Button onPress={sendMailAsync} title="Send birthday wishes" />
+      <Button
+        onPress={() =>
+          MailComposer.openClientAsync({
+            title: 'Which email app do you use?',
+            cancelLabel: 'Cancel',
+          })
+        }
+        title="Open client"
+      />
+      <Button onPress={sendMailAsync} title="Send birthday wishes" style={styles.sendMailButton} />
       {status && <MonoText>Status: {status}</MonoText>}
     </View>
   );
 }
+
+MailComposerScreen.navigationOptions = {
+  title: 'MailComposer',
+};
 
 const styles = StyleSheet.create({
   container: {
@@ -72,5 +53,8 @@ const styles = StyleSheet.create({
   capabilitiesContainer: {
     alignItems: 'stretch',
     paddingBottom: 20,
+  },
+  sendMailButton: {
+    marginTop: 10,
   },
 });