+++ packages/expo-mail-composer/src/ExpoMailComposer.web.ts
@@ -1,4 +1,9 @@
-import { MailComposerOptions, MailComposerResult, MailComposerStatus } from './MailComposer.types';
+import {
+  MailClientOptions,
+  MailComposerOptions,
+  MailComposerResult,
+  MailComposerStatus,
+} from './MailComposer.types';
 
 function removeNullishValues(obj) {
   for (const propName in obj) {
@@ -19,6 +24,7 @@ function checkValue(value?: string[] | string): string | null {
 }
 
 export default {
+  async openClientAsync(options: MailClientOptions): Promise<void> {},
   async composeAsync(options: MailComposerOptions): Promise<MailComposerResult> {
     if (typeof window === 'undefined') {
       return { status: MailComposerStatus.CANCELLED };